# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lang Cards API v1",
        "description": "Lang Cards API ",
        "contact": {
            "email": "vlad2000100600@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/setInviter": {
            "post": {
                "tags": [
                    "Привязка к пользователю, который пригласил в систему"
                ],
                "summary": "Установить пригласившего пользователя",
                "description": "Позволяет авторизованному пользователю указать другого пользователя, который его пригласил, по коду приглашения.",
                "operationId": "setInviter",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptLanguageHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetInviterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная установка пригласившего пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Благодарим! Вы указали пригласившего пользователя."
                                        },
                                        "data": {
                                            "required": [
                                                "inviter"
                                            ],
                                            "properties": {
                                                "inviter": {
                                                    "$ref": "#/components/schemas/InviterResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "description": "Не найден пользователь с таким кодом",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Не найден пользователь, владеющий данным invite-кодом"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Конфликт — код уже установлен, код невалиден или нарушает бизнес-логику",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Конфликт — код уже установлен, код невалиден или нарушает бизнес-логику"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "invite_code": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invite Code is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/timezones": {
            "get": {
                "tags": [
                    "Временные зоны"
                ],
                "summary": "Получить список часовых поясов с пагинацией и фильтрацией",
                "operationId": "getTimezones",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptLanguageHeader"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы (целое число >= 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "countOnPage",
                        "in": "query",
                        "description": "Количество элементов на странице (целое число >= 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Список полей через запятую, которые нужно вернуть",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными и пагинацией",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Данные часовых поясов"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/TimezoneResource"
                                                    }
                                                },
                                                "pagination": {
                                                    "description": "Информация о пагинации",
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 100
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/NotVerifiedEmail"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SetInviterRequest": {
                "title": "Set Inviter Request (Установка пригласившего пользователя)",
                "description": "Запрос для установки пригласившего пользователя по коду приглашения. Код должен состоять из 16 символов: только заглавные латинские буквы и цифры (A–Z, 0–9).",
                "required": [
                    "invite_code"
                ],
                "properties": {
                    "invite_code": {
                        "description": "Пригласительный код. Должен содержать ровно 16 символов: заглавные латинские буквы и цифры.",
                        "type": "string",
                        "maxLength": 16,
                        "minLength": 16,
                        "pattern": "^[A-Z0-9]{16}$",
                        "example": "AB12CD34EF56GH78"
                    }
                },
                "type": "object"
            },
...
            "InviterResource": {
                "title": "Inviter Resource",
                "description": "Информация о пригласившем пользователе",
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор пользователя",
                        "type": "integer",
                        "example": 123
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "avatar": {
                        "description": "URL аватара пользователя, может отсутствовать",
                        "type": "string",
                        "example": "https://example.com/avatars/123.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TimezoneResource": {
                "title": "Timezone Resource (Ресурс для получения информации о временных зонах, поддерживаемых в системе)",
                "description": "Ресурс часового пояса",
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Название часового пояса",
                        "type": "string"
                    },
                    "offset_utc": {
                        "description": "Смещение от UTC в минутах",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Дата и время создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "NotVerifiedEmail": {
                "description": "Электронный адрес пользователя не подтвержден",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Электронная почта авторизованного пользователя не подтверждена"
                                },
                                "errors": {
                                    "type": "object",
                                    "example": null,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Пользователь не авторизован",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Пользователь не авторизован и не имеет доступа к данным"
                                },
                                "errors": {
                                    "type": "object",
                                    "example": null,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "AcceptLanguageHeader": {
                "name": "Accept-Language",
                "in": "header",
                "description": "Язык локали",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "ru",
                    "enum": [
                        "ru",
                        "en"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}